testsuite: track main's builtin renames and the 64-bit instruction buffer
main renamed __builtin_rvtt_sfpsetcc_v to __builtin_rvtt_sfpsetcc (the _v
suffix disappeared when the vector/scalar split was folded into the mod
TYPE field) and widened the instruction buffer to unsigned long.  Update
every craq-added test to the post-rebase spelling: 194 sfpsetcc sites
across 150 files, 47 buffer declarations.

+++ gcc/testsuite/g++.target/riscv/tt/sfpi/setexp-fold-bh.C
@@ -3,7 +3,7 @@
 // { dg-options "-mcpu=tt-bh-tensix -O2 -I [SFPI]/include -fno-exceptions -fno-rtti -mtt-tensix-optimize-setexp-fold" }
 
 namespace ckernel{
-  extern volatile unsigned instrn_buffer[];
+  extern volatile unsigned long instrn_buffer[];
 }
 #include <sfpi.h>
